Fix style + refactor add choice Component

+++ src/components/Choice/Add.vue
@@ -2,7 +2,7 @@
   <div class="container">
     <div v-if="orders.length">
       <div class="mt-3 d-flex flex-row justify-content-center">
-        <div v-for="(order, key) in orders" :key="key" @click="orderSelected(order)" class="card border-warning mx-2 my-2">
+        <div v-for="(order, key) in orders" :key="key" class="card border-warning mx-2 my-2">
           <div class="card-header text-center">
             <h2>{{ order.vendor.name }}</h2>
             <small class="badge badge-warning">
@@ -10,24 +10,16 @@
             </small>
           </div>
           <div class="card-body text-dark">
-            <ProductsInputList :products="order.vendor.products" :order="order" @productSelected="productSelected"></ProductsInputList>
+            <ProductsInputList :order.sync="order" @update:order="setChoice"></ProductsInputList>
           </div>
           <div class="card-footer bg-transparent border-warning">
-            <div>
-              <strong>Deadline at:</strong>
-              <small>{{ order.deadlineAt | moment }}</small>
-            </div>
-            <div>
-              <strong>Choice:</strong>
-              <small v-if="order.choice && order.choice.id">{{ order.choice.product.name }}</small>
-            </div>
             <div v-if="order.choice && order.choice.comment">
               <strong>Comment:</strong>
-              <small v-if="order.choice">{{ order.choice.comment }}</small>
+              <small>{{ order.choice.comment }}</small>
             </div>
             <div class="text-center mt-1">
               <button
-                v-if="order.choice.product && beforeDeadline(order)"
+                v-if="order.choice && order.choice.product && beforeDeadline(order)"
                 v-on:click="removeChoice(order)"
                 class="btn btn-sm btn-danger col-white"
               >
@@ -40,11 +32,11 @@
       <div class="row justify-content-center mt-4">
         <div class="col-sm-6">
           <span>Comment:</span>
-          <textarea v-model="choice.comment" class="form-control border-warning"></textarea>
+          <textarea v-model="comment" class="form-control border-warning"></textarea>
         </div>
       </div>
       <div class="d-flex justify-content-center mt-4">
-        <button :disabled="!choice.product" class="btn btn-lg btn-warning col-white" v-on:click="sendForm">Send</button>
+        <button :disabled="!isProductSelected" class="btn btn-lg btn-warning col-white" v-on:click="sendForm">Send</button>
       </div>
     </div>
   </div>
@@ -61,48 +53,49 @@ export default {
   data: () => {
     return {
       orders: [],
-      choice: {
-        order: {},
-        product: null,
-        comment: ''
-      },
+      choice: null,
       service: new ChoiceService(),
       orderService: new OrderService()
     };
   },
   methods: {
-    sendForm: async function() {
+    sendForm: function() {
       if (this.choice.product) {
-        const choice = await this.service.store(this.$store.getters.user, this.choice.order, this.choice.product, this.choice.comment);
-        this.orders.find(x => x.id === choice.orderId).choice = {
-          id: choice.id,
-          product: this.choice.product,
-          comment: this.choice.comment
-        };
-        this.choice.comment = '';
+        this.service.store(this.$store.getters.user, this.choice.order, this.choice.product, this.choice.comment).then(choice => {
+          this.orders.find(x => x.id === choice.orderId).choice = choice;
+          this.choice = null;
+        });
       }
     },
     removeChoice: async function(order) {
-      await this.service.remove(order.id, order.choice.id);
-      this.choice.comment = '';
-      this.orders.find(x => x.id === order.id).choice = {
-        id: null,
-        product: null,
-        comment: ''
-      };
-    },
-    productSelected: function(product, order) {
-      this.orders.find(x => x.id === order.id).choice.product = product;
-      this.choice.product = product;
-    },
-    orderSelected: function(order) {
-      this.choice.order = order;
+      if (order.id && order.choice.id) {
+        await this.service.remove(order.id, order.choice.id);
+      }
+      this.choice = null;
+      this.orders.find(x => x.id === order.id).choice = null;
     },
     beforeDeadline: function(order) {
       return new Date(order.deadlineAt) > new Date();
+    },
+    setChoice(order) {
+      this.choice = order.choice;
+      this.choice.order = order;
     }
   },
-  computed: {},
+  computed: {
+    comment: {
+      get() {
+        return this.choice && this.choice.comment;
+      },
+      set(comment) {
+        const choice = this.choice || {};
+        this.choice = Object.assign(choice, { comment });
+      }
+    },
+    isProductSelected() {
+      return this.choice && this.choice.product;
+    }
+  },
   async created() {
     this.orders = await this.orderService.getAllWithProductChoices();
   },
